pyproject.toml
[tool.poetry]
name = "ecoshop"
version = "0.1.0"
description = ""
authors = ["alexandrurentsev <alexandrurentsev@yandex.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.0"
uvicorn = "^0.27.0.post1"
pydantic-settings = "^2.1.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
asyncpg = "^0.29.0"
alembic = "^1.13.1"
passlib = "^1.7.4"
python-jose = "^3.3.0"
pydantic = {extras = ["email"], version = "^2.6.1"}
bcrypt = "^4.1.2"
jinja2 = "^3.1.3"
python-multipart = "^0.0.9"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
celery = "^5.3.6"
flower = "^2.0.1"
pillow = "^10.2.0"
sqladmin = "^0.16.1"
itsdangerous = "^2.1.2"
pytest = "^8.0.2"


[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
flake8 = "^7.0.0"
autopep8 = "^2.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
